refactor: editor change handler
